Hw04 auth

+++ services/contacts.js
@@ -1,26 +1,48 @@
 const { Contact } = require("../models");
 
-const listContacts = async () => {
-  const results = await Contact.find();
+const listContacts = async (pagination, filter) => {
+  const { page, limit } = pagination;
+  const skip = page * limit - limit;
+  const results = await Contact.find(filter, "id name email phone favorite", {
+    skip,
+    limit: Number(limit),
+  });
   return results;
 };
 
-const getContactById = async (contactId) => {
-  const results = await Contact.findOne({ _id: contactId });
+const getContactById = async (contactId, ownerId) => {
+  const filter = { _id: contactId, owner: ownerId };
+  const results = await Contact.find(filter);
   return results;
 };
 
-const removeContact = async (contactId) => {
+const removeContact = async (contactId, ownerId) => {
+  const filter = { _id: contactId, owner: ownerId };
+  const user = await Contact.findById(filter);
+
+  if (!user || user.length < 1) return false;
+
   const result = await Contact.findByIdAndDelete({ _id: contactId });
   return result;
 };
 
 const addContact = async (body) => {
+  const user = await Contact.find({ email: body.email }, { new: true });
+
+  if (user.length > 0) {
+    return false;
+  }
+
   const result = await Contact.create(body);
   return result;
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, ownerId) => {
+  const filter = { _id: contactId, owner: ownerId };
+  const user = await Contact.findById(filter);
+
+  if (!user || user.length < 1) return false;
+
   const result = await Contact.findByIdAndUpdate(
     { _id: contactId },
     { ...body },
@@ -28,7 +50,6 @@ const updateContact = async (contactId, body) => {
   );
   return result;
 };
-
 
 module.exports = {
   listContacts,
